Move object equality functions to another package for clarity
